pyproject.toml
[tool.poetry]
name = "FileETL"
version = "0.1.0"
description = "Project to unify different sources in one output"
authors = ["JangasCodingplace <janisgoesser92@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7.1"
python-dotenv = "^0.17.1"
pandas = "^1.2.4"
langdetect = "^1.0.9"

[tool.poetry.dev-dependencies]
flake8 = "^3.9.2"
pytest = "^6.2.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
